intelliclaim-vision-backend/requirements.txt
# =============================================================================
# AI Insurance Processing Application - Complete Requirements
# Compatible with Python 3.13+
# =============================================================================

# Core FastAPI and Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.10
starlette>=0.40.0

# Authentication and Security
pyjwt>=2.8.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Database and ORM
sqlalchemy>=2.0.30
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Caching and Queue Systems
redis>=5.0.0
aioredis>=2.0.1
celery>=5.3.0
kombu>=5.3.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=43.0.0

# OpenAI and AI Services
openai>=1.50.0
tiktoken>=0.7.0
google-generativeai>=0.8.3

# Computer Vision (Python 3.13 compatible)
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.206
torch>=2.1.0
torchvision>=0.16.0
scikit-image>=0.22.0
imageio>=2.31.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.22
openpyxl>=3.1.0
xlrd>=2.0.1
pytesseract>=0.3.10

# Vector Databases
chromadb>=0.4.15
pinecone-client>=2.2.4
sentence-transformers>=2.2.2

# HTTP Client and Web Utilities
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.31.0

# Configuration and Environment
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.0

# Monitoring and Logging
loguru>=0.7.0
sentry-sdk[fastapi]>=2.0.0
prometheus-client>=0.19.0

# Data Processing and Analytics
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Date and Time Utilities
python-dateutil>=2.8.0
pytz>=2023.0

# File Storage and Cloud Services
boto3>=1.34.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.19.0

# Email Services
fastapi-mail>=1.4.0

# Background Processing and Scheduling
dramatiq>=1.15.0
apscheduler>=3.10.0

# Data Validation
marshmallow>=3.20.0
cerberus>=1.3.5

# Development Tools and Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# Testing Framework
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0
factory-boy>=3.3.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Performance and Profiling
memory-profiler>=0.61.0

# Additional Utilities
tqdm>=4.60.0
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
jinja2>=3.1.0

# Template and UI
fastapi-users>=12.0.0
fastapi-users[sqlalchemy]>=12.0.0
